requirements.txt
Flask==3.1.1
gunicorn==23.0.0
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Werkzeug==3.1.3
SQLAlchemy==2.0.19
psycopg2-binary==2.9.9
